# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/patients": {
      "post": {
        "operationId": "PatientsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePatientDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or revoked token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but lacking the permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Create a patient file; the file number is allocated here",
        "tags": [
          "patients"
        ]
      },
      "get": {
        "operationId": "PatientsController_search",
        "parameters": [
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "Name or file number; tolerant of misspelling",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "LEAD",
                "SCHEDULED",
                "PRE_OP",
                "POST_OP",
                "FOLLOW_UP",
                "DISCHARGED",
                "INACTIVE"
              ]
            }
          },
          {
            "name": "country",
            "required": false,
            "in": "query",
            "description": "ISO 3166-1 alpha-2",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "assignedDoctorId",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "procedure",
            "required": false,
            "in": "query",
            "description": "Matches the procedure name of a recorded surgery",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "surgeryFrom",
            "required": false,
            "in": "query",
            "description": "Surgeries on or after this date",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "surgeryTo",
            "required": false,
            "in": "query",
            "description": "Surgeries on or before this date",
            "schema": {
              "format": "date-time",
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "description": "Id of the last row from the previous page",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "maximum": 100,
              "default": 25,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientPageDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or revoked token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but lacking the permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Search within the caller’s scope",
        "tags": [
          "patients"
        ]
      }
    },
    "/patients/{id}": {
      "get": {
        "operationId": "PatientsController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or revoked token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but lacking the permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or outside the caller’s scope — deliberately indistinguishable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "One patient file",
        "tags": [
          "patients"
        ]
      },
      "patch": {
        "operationId": "PatientsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePatientDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PatientDto"
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or revoked token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but lacking the permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or outside the caller’s scope — deliberately indistinguishable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Update demographics and status",
        "tags": [
          "patients"
        ]
      },
      "delete": {
        "operationId": "PatientsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deactivated"
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or revoked token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but lacking the permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or outside the caller’s scope — deliberately indistinguishable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Deactivate a file — soft delete; records are retained by law",
        "tags": [
          "patients"
        ]
      }
    },
    "/patients/{id}/medical-profile": {
      "put": {
        "operationId": "PatientsController_upsertMedicalProfile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MedicalProfileDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Stored"
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or revoked token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but lacking the permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or outside the caller’s scope — deliberately indistinguishable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Allergies, chronic conditions, smoking and alcohol",
        "tags": [
          "patients"
        ]
      }
    },
    "/patients/{id}/assignments": {
      "get": {
        "operationId": "PatientsController_listAssignments",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AssignmentDto"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or revoked token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but lacking the permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or outside the caller’s scope — deliberately indistinguishable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Staff currently assigned to this patient",
        "tags": [
          "patients"
        ]
      },
      "post": {
        "operationId": "PatientsController_assign",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignStaffDto"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Assigned"
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or revoked token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but lacking the permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or outside the caller’s scope — deliberately indistinguishable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Assign a staff member — this decides who can see the file",
        "tags": [
          "patients"
        ]
      }
    },
    "/patients/{id}/assignments/{staffId}": {
      "delete": {
        "operationId": "PatientsController_unassign",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "staffId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Assignment ended"
          },
          "400": {
            "description": "Validation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or revoked token",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Authenticated but lacking the permission",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "Not found, or outside the caller’s scope — deliberately indistinguishable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "End an assignment",
        "tags": [
          "patients"
        ]
      }
    }
  },
  "info": {
    "title": "Klinik Takip API",
    "description": "Doctor–patient follow-up platform. Every endpoint requires a bearer token unless documented otherwise; permissions are checked per route and patient access is scoped per role.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ErrorResponseDto": {
        "type": "object",
        "properties": {
          "statusCode": {
            "type": "number",
            "example": 400
          },
          "message": {
            "type": "object",
            "description": "Human-readable text, or a machine-readable code for authentication failures (INVALID_CREDENTIALS, MFA_REQUIRED, ACCOUNT_LOCKED, …).",
            "example": "Missing permission: patients.write"
          },
          "error": {
            "type": "string",
            "example": "Forbidden"
          }
        },
        "required": [
          "statusCode",
          "message"
        ]
      },
...
      "CreatePatientDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "format": "date"
          },
          "sex": {
            "type": "string",
            "enum": [
              "FEMALE",
              "MALE",
              "OTHER",
              "UNDISCLOSED"
            ]
          },
          "country": {
            "type": "string",
            "description": "ISO 3166-1 alpha-2, drives language and discharge advice"
          },
          "city": {
            "type": "string"
          },
          "nationality": {
            "type": "string"
          },
          "preferredLanguage": {
            "type": "string",
            "default": "tr"
          },
          "referralSource": {
            "type": "string",
            "description": "Instagram, Google, agency, referral…"
          },
          "assignedDoctorId": {
            "type": "string"
          },
          "agencyId": {
            "type": "string"
          }
        },
        "required": [
          "firstName",
          "lastName",
          "birthDate",
          "sex",
          "country"
        ]
      },
      "PatientDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "mrn": {
            "type": "string",
            "description": "File number, e.g. 2026-K7RMPX",
            "example": "2026-K7RMPX"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "format": "date"
          },
          "sex": {
            "type": "string",
            "enum": [
              "FEMALE",
              "MALE",
              "OTHER",
              "UNDISCLOSED"
            ]
          },
          "country": {
            "type": "string",
            "description": "ISO 3166-1 alpha-2, always upper case",
            "example": "DE"
          },
          "city": {
            "type": "object",
            "nullable": true
          },
          "nationality": {
            "type": "object",
            "nullable": true
          },
          "preferredLanguage": {
            "type": "string",
            "example": "tr"
          },
          "referralSource": {
            "type": "object",
            "nullable": true,
            "description": "Instagram, Google, agency, referral…"
          },
          "status": {
            "type": "string",
            "enum": [
              "LEAD",
              "SCHEDULED",
              "PRE_OP",
              "POST_OP",
              "FOLLOW_UP",
              "DISCHARGED",
              "INACTIVE"
            ]
          },
          "assignedDoctorId": {
            "type": "object",
            "format": "uuid",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true,
            "description": "Set when the file is deactivated. Records are retained by law, not destroyed."
          }
        },
        "required": [
          "id",
          "mrn",
          "firstName",
          "lastName",
          "birthDate",
          "sex",
          "country",
          "city",
          "nationality",
          "preferredLanguage",
          "referralSource",
          "status",
          "assignedDoctorId",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "PatientPageDto": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PatientDto"
            }
          },
          "nextCursor": {
            "type": "object",
            "nullable": true,
            "description": "Pass as `cursor` for the next page. Null on the last page."
          }
        },
        "required": [
          "items",
          "nextCursor"
        ]
      },
      "UpdatePatientDto": {
        "type": "object",
        "properties": {
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "birthDate": {
            "type": "string",
            "format": "date"
          },
          "sex": {
            "type": "string",
            "enum": [
              "FEMALE",
              "MALE",
              "OTHER",
              "UNDISCLOSED"
            ]
          },
          "country": {
            "type": "string"
          },
          "city": {
            "type": "string"
          },
          "preferredLanguage": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "LEAD",
              "SCHEDULED",
              "PRE_OP",
              "POST_OP",
              "FOLLOW_UP",
              "DISCHARGED",
              "INACTIVE"
            ]
          },
          "assignedDoctorId": {
            "type": "string"
          }
        }
      },
      "MedicalProfileDto": {
        "type": "object",
        "properties": {
          "bloodType": {
            "type": "string"
          },
          "allergies": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "chronicConditions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "currentMedications": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "smoking": {
            "type": "boolean"
          },
          "alcohol": {
            "type": "boolean"
          },
          "notes": {
            "type": "string"
          }
        }
      },
      "AssignedStaffDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "firstName": {
            "type": "string"
          },
          "lastName": {
            "type": "string"
          },
          "title": {
            "type": "object",
            "nullable": true
          }
        },
        "required": [
          "id",
          "firstName",
          "lastName",
          "title"
        ]
      },
      "AssignmentDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "role": {
            "type": "string",
            "enum": [
              "SUPER_ADMIN",
              "DOCTOR",
              "NURSE",
              "COORDINATOR",
              "FINANCE",
              "PATIENT",
              "CAREGIVER"
            ]
          },
          "assignedAt": {
            "type": "string",
            "format": "date-time"
          },
          "staff": {
            "$ref": "#/components/schemas/AssignedStaffDto"
          }
        },
        "required": [
          "id",
          "role",
          "assignedAt",
          "staff"
        ]
      },
      "AssignStaffDto": {
        "type": "object",
        "properties": {
          "staffId": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "enum": [
              "SUPER_ADMIN",
              "DOCTOR",
              "NURSE",
              "COORDINATOR",
              "FINANCE",
              "PATIENT",
              "CAREGIVER"
            ]
          }
        },
        "required": [
          "staffId",
          "role"
        ]
      }
    }
  }
}